+++ tests/testthat/test-get_naturalearth.R
@@ -1,9 +1,32 @@
 test_that("get_coastline returns expected objects", {
-  bf <- BirdFlowModels::amewoo
+  bf <- new_BirdFlow()
+  bf$geom$crs <- terra::crs(paste0("+proj=moll +lon_0=-90 +x_0=0 +y_0=0",
+                            " +ellps=WGS84 +units=m +no_defs"))
+  bf$geom$ext <- c(-1564958.2138293, 2514826.55772382,
+                      2822153.67510416, 6421966.58574292)
+
+  # Using "new" method
   expect_s3_class(coast <- get_coastline(bf), class = c("sf", "data.frame") )
   expect_s3_class(coast$geometry, c("sfc_GEOMETRY", "sfc") )
   expect_s3_class(coast$geometry[1],  c("sfc_LINESTRING", "sfc"))
   expect_true( sf::st_crs(coast) == sf::st_crs(crs(bf) ) )
+  expect_equal(nrow(coast), 269 )
+
+  # using "old" method
+  expect_no_error(coast2 <- get_naturalearth(bf, type = "coastline",
+                            force_old_method = TRUE) )
+  expect_s3_class(coast2, c("sf", "data.frame"))
+  expect_s3_class(coast2$geometry, c("sfc_GEOMETRY", "sfc") )
+  expect_s3_class(coast2$geometry[1],  c("sfc_LINESTRING", "sfc"))
+  expect_true( sf::st_crs(coast2) == sf::st_crs(crs(bf) ) )
+  expect_equal(nrow(coast2), 231 )
+
+  if(interactive()){
+    # Due to buffering and reprojection issues the extent of the
+    # two versions is slightly different
+    plot(coast2)
+    plot(coast, add = TRUE, col = rgb(0, 0, 1, .2), lwd = 4)
+  }
 
 })
 
@@ -61,8 +84,8 @@ test_that("get_naturalearth works with non-default scale", {
   bf <- BirdFlowModels::amewoo
 
   expect_no_error(
-  grat_med <- get_naturalearth(bf, type = "graticules_30",
-                               category = "physical", scale = "large")
+    grat_med <- get_naturalearth(bf, type = "graticules_30",
+                                 category = "physical", scale = "large")
   )
 
 })
@@ -94,6 +117,122 @@ test_that("get_naturalearth() works at edge of WGS84", {
   expect_snapshot(coast) # eastern Australia and islands
 
 })
+
+
+
+test_that("get_naturalearth() works at edge of WGS84 with old method", {
+
+  # Construct a psuedo BirdFlow object that has a crs centered on the edge
+  # of the wgs84 projection (used by rnaturalearth)
+  seam_crs <- crs("+proj=moll +lon_0=180 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs +type=crs")  # mollweide centered on 180 deg lon.
+  bf <- new_BirdFlow()
+  bf$geom$crs <- seam_crs
+  bf$geom$ext <- c(-2000000, 2000000,-5000000, 1000000)
+
+  # Visualize what we are hoping to do
+  if(FALSE){ # This code is to run manually while debugging
+    all_coast <- rnaturalearth::ne_coastline(returnclass = "sf")
+    coast_proj <- sf::st_transform(all_coast, seam_crs)
+    op <- par( no.readonly = TRUE)
+    par(oma = rep(0, 4), mar = rep(0, 4))
+    plot(coast_proj[, "geometry"])
+    abline(v = c(-2000000, 2000000), col = "blue")
+    abline(h = c(-5000000, 1000000), col = "red")
+
+    par(op)
+  }  # end skip visualization
+
+  expect_no_error(
+    coast <- get_naturalearth(bf, type = "coastline",
+                              scale = 110, force_old_method = TRUE) )
+  if(interactive()){
+    plot(coast)
+  }
+
+  expect_s3_class(coast, c("sf", "data.frame"))
+  expect_s3_class(coast$geometry, c("sfc_GEOMETRY", "sfc") )
+  expect_s3_class(coast$geometry[1],  c("sfc_LINESTRING", "sfc"))
+  expect_true( sf::st_crs(coast) == sf::st_crs(crs(bf) ) )
+  expect_equal(nrow(coast), 24 )
+
+})
+
+test_that("get_naturalearth() works with mollweide and broken bounding box",{
+  # Construct a psuedo BirdFlow object with the extent and
+  #  projection that a user submitted.  This is a mollweide where
+  # one corner of the bounding box is not on the map.
+
+  bf <- new_BirdFlow()
+  bf$geom$crs <- crs(paste0("+proj=moll +lon_0=-90 +x_0=0 +y_0=0 +datum=WGS84",
+  " +units=m +no_defs"))
+  bf$geom$ext <- c(-12400447.5244956, 7148844.12665357,
+                   3241411.87373662, 7978416.80488446)
+
+
+  if(interactive()){
+    # This failed during with the back transform extent method
+    coast <- rnaturalearth::ne_coastline(returnclass = "sf") |>
+      sf::st_transform( crs = crs(bf))
+
+    plot(coast[ , "geometry"])
+    plot(terra::ext(bf), add = TRUE, border = "red")
+  }
+
+  expect_no_error(coast2 <- get_coastline(bf))
+  expect_equal(nrow(coast2), 668 )
+})
+
+
+
+test_that("get_naturalearth() works with lambert equal area (laea)",{
+  # Construct a psuedo BirdFlow object
+  bf <- new_BirdFlow()
+  bf$geom$crs <- crs(paste0("+proj=laea +lat_0=39.161 +lon_0=-85.094 +x_0=0 ",
+                            "+y_0=0 +datum=WGS84 +units=m +no_defs"))
+  bf$geom$ext <- c(-2410760.5, 1958109.5, -1548178, 1658294)
+
+  expect_no_error(coast2 <- get_coastline(bf))
+  expect_equal(nrow(coast2),  241 )  # original run through had 134
+
+  if(interactive()){
+    coast <- rnaturalearth::ne_coastline(returnclass = "sf") |>
+      sf::st_transform( crs = crs(bf))
+    plot(coast[ , "geometry"])
+    plot(terra::ext(bf), add = TRUE, border = "red")
+  }
+
+})
+
+test_that("get_naturalearth() issues appropriate warning with empty extent", {
+
+  crs <- crs("+proj=moll +lon_0=180 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs +type=crs")  # mollweide centered on 180 deg lon.
+  bf <- new_BirdFlow()
+  bf$geom$crs <- crs
+  bf$geom$ext <- c(-500000, 1000000, 1000000, 2000000)
+
+  # Visualize empty extent in target projection
+  if(interactive()){
+    all_coast <- rnaturalearth::ne_coastline(returnclass = "sf")
+    coast_proj <- sf::st_transform(all_coast, crs)
+    op <- par( no.readonly = TRUE)
+    par(oma = rep(0, 4), mar = rep(0, 4))
+    plot(coast_proj[, "geometry"])
+    terra::plot(terra::ext(bf), add = TRUE, border = "red")
+    par(op)
+  }
+
+  # new method
+  expect_warning( a <- get_coastline(bf, buffer = 0),
+                  "No objects within extent. Returning empty sf object.")
+
+  # old method
+  expect_warning( a <- get_naturalearth(bf, "coastline",  buffer = 0,
+                                        force_old_method = TRUE),
+                  "No objects within extent. Returning empty sf object.")
+
+
+})
+